requirements.txt
Flask~=3.1.1
ipaddress~=1.0.23
httpagentparser~=1.9.5
requests~=2.32.4
hpack~=4.1.0
hyperframe~=6.1.0
h11~=0.16.0
cryptography~=45.0.4
pip~=25.0.1
rsa~=4.9.1
pyasn1~=0.6.1
typing_extensions~=4.14.0
sniffio~=1.3.1
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.1.8
Jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.4.26
urllib3~=2.5.0
packaging~=25.0
cffi~=1.17.1
google-api-python-client~=2.173.0
h2~=4.2.0
greenlet~=3.2.3
gunicorn~=23.0.0
anyio~=4.9.0
pyparsing~=3.2.3
httplib2~=0.22.0
pycparser~=2.22
SQLAlchemy~=2.0.41
uritemplate~=4.2.0
CacheControl~=0.14.3
msgpack~=1.1.1
PyJWT~=2.10.1